chore: Upgrade dependencies and move stuff in vision module
