+++ Cargo.toml
@@ -16,8 +16,8 @@ path = "src/main.rs"
 
 [dependencies]
 bevy = { version = "0.16.0", features = ["dynamic_linking"] }
-bevy_ecs_tilemap = {git = "https://github.com/miketwenty1/bevy_ecs_tilemap.git", branch = "update-bevy-16.0-release" }
-bevy_asset_loader = { git = "https://github.com/NiklasEi/bevy_asset_loader.git", features = ["2d"], branch = "bevy_main" }
+bevy_ecs_tilemap = { version = "0.16" }
+bevy_asset_loader = { version = "0.23", features = ["2d"] }
 bevy_behave = { version = "0.3" }
 accesskit = "0.17"
 avian2d = { git = "https://github.com/Jondolf/avian.git", branch = "main" }
